pyproject.toml
[project]
name = "project4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.4",
    "django-bootstrap5>=25.1",
    "django-filter>=25.1",
    "django-vite>=3.1.0",
    "gunicorn>=23.0.0",
    "inertia-django>=1.2.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "rollbar>=1.3.0",
    "ruff>=0.12.1",
    "whitenoise>=6.9.0",
]

[build-system] 
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel] 
packages = ["task_manager"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "task_manager.settings"
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]
